Refactor theme observer
