+++ app/javascript/utils/debounce.ts
@@ -1,10 +1,13 @@
-export const debounce = (
-  func: (...args: unknown[]) => void,
+/* eslint-disable @typescript-eslint/no-explicit-any */
+// any is descriptive here
+// using unknown would make thigs very hard to read
+export function debounce<F extends (...args: any[]) => any>(
+  func: F,
   delay: number
-): ((...args: unknown[]) => void) => {
-  let inDebounce: number | null
-  return (...args: unknown[]) => {
-    window.clearTimeout(inDebounce as number)
-    inDebounce = window.setTimeout(() => func(...args), delay)
+): (...args: Parameters<F>) => void {
+  let inDebounce: ReturnType<typeof setTimeout> | null = null
+  return (...args: Parameters<F>): void => {
+    clearTimeout(inDebounce as ReturnType<typeof setTimeout>)
+    inDebounce = setTimeout(() => func(...args), delay)
   }
 }
