+++ app/javascript/utils/index.ts
@@ -8,3 +8,4 @@ export { scrollIntoView } from './scroll-into-view'
 export { redirectTo } from './redirect-to'
 export { typecheck } from './typecheck'
 export { showSiteFooterOnTurboLoad } from './show-site-footer'
+export { debounce } from './debounce'
